refactored dropdown functionality into voteitem and delete dropdown component
